feat(native): enforce wire-version handshake on native protocol connections
The native server now performs the Hello/HelloAck exchange as the first
step of every TCP connection via `perform_server_handshake`, storing the
negotiated version on `NativeSession`.  Servers that receive a client
whose version range doesn't overlap respond with a `HelloErrorFrame` and
close the connection.

On the client side, `NativeConnection::connect` and `connect_tls` run
the handshake immediately after the TCP/TLS connection is established.
The client reads the 4-byte magic prefix to distinguish a `HelloAckFrame`
from a `HelloErrorFrame` and surfaces version mismatches as a typed
`NodeDbError::HandshakeFailed`.

Per-operation limit enforcement (`check_op_limits`) is now also called in
`direct_ops` before quota checks, and the `LimitExceeded` variant is wired
into the server `Error` enum so it converts cleanly to `NodeDbError::BadRequest`.

+++ nodedb-client/src/native/client.rs
@@ -218,7 +218,7 @@ impl NodeDb for NativeClient {
                 OpCode::GraphHop,
                 TextFields {
                     start_node: Some(start.as_str().to_string()),
-                    depth: Some(depth as u64),
+                    depth: Some(depth as u32),
                     edge_label: edge_filter.and_then(|f| f.labels.first().cloned()),
                     ..Default::default()
                 },
